Proyecto-V0.1.8-GranActualizacion,UpdateUser,Minimos2,Exit,Loggin1Vez,Inventario

+++ app/src/main/java/dsa/proyectoandroid/g6/models/Product.java
@@ -3,13 +3,13 @@ package dsa.proyectoandroid.g6.models;
 public class Product {
     private String id;
     private String nombre;
-    private Double precio;
+    private Integer precio;
 
     //Constructors
     public Product() {
     }
 
-    public Product(String id, String nombre, Double precio) {
+    public Product(String id, String nombre, Integer precio) {
         this.id = id;
         this.nombre = nombre;
         this.precio = precio;
@@ -32,11 +32,11 @@ public class Product {
         this.nombre = nombre;
     }
 
-    public Double getPrecio() {
+    public Integer getPrecio() {
         return precio;
     }
 
-    public void setPrecio(Double precio) {
+    public void setPrecio(Integer precio) {
         this.precio = precio;
     }
 }
